# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Momentum Dashboard API
  version: 0.7.0
  summary: OpenAPI-first contracts for React + FastAPI app (Parquet+SQLite).
  description: Contract-first spec. Clients must be generated from this file. All
    endpoints are local-only and authenticated as N/A for now.
servers:
- url: http://localhost:8000/api/v1
...
paths:
...
  /api/v1/instruments:
    get:
      tags:
      - Instruments
      operationId: listInstruments
      summary: Search/autocomplete instruments
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Instruments
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UniverseInstrument'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /api/v1/instruments/{symbol}/detail:
    get:
      tags:
      - Instruments
      summary: Get instrument drawer detail (backend-only)
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
        description: Canonical symbol, e.g. TCS.NS
      - name: run_id
        in: query
        required: false
        schema:
          type: string
        description: Use a specific snapshot; defaults to pinned or latest
      responses:
        '200':
          description: Drawer detail payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrawerDetail'
              examples:
                default:
                  $ref: '#/components/examples/DrawerDetailExample'
...
  /api/v1/momentum/heatmap:
    get:
      tags:
      - Momentum
      summary: NSE sector and industry momentum heatmap snapshot
      operationId: getMomentumHeatmap
      parameters:
      - $ref: '#/components/parameters/IfNoneMatch'
      - in: query
        name: as_of
        required: false
        description: >
          Optional trading date (YYYY-MM-DD, IST) to retrieve a historical snapshot; defaults to the latest available run.
        schema:
          type: string
          format: date
      - in: query
        name: include_industries
        required: false
        schema:
          type: boolean
          default: false
        description: Include industry-level breakdown nested within each sector.
      - in: query
        name: include_constituents
        required: false
        schema:
          type: boolean
          default: false
        description: Include top/bottom constituent momentum summaries for each node.
      responses:
        '200':
          description: Momentum heatmap snapshot
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MomentumHeatmapResponse'
              examples:
                example:
                  value:
                    as_of: '2025-10-12T10:45:00+05:30'
                    trade_date: '2025-10-10'
                    session: cash
                    run_id: '20251012104500'
                    source: nse
                    latency_sec: 120
                    sectors:
                    - name: NIFTY IT
                      symbol: NIFTYIT
                      change_1d: 1.85
                      change_1w: 3.12
                      change_1m: 7.9
                      momentum_score: 0.82
                      turnover_ratio: 1.4
                      total_turnover_cr: 3250.7
                      last_updated: '2025-10-12T10:43:00+05:30'
                      advance_decline:
                        advancers: 18
                        decliners: 4
                        unchanged: 0
                      industries:
                      - name: IT Services
                        symbol: NSEITSE
                        change_1d: 1.9
                        change_1w: 3.5
                        change_1m: 8.4
                        momentum_score: 0.85
                        turnover_ratio: 1.5
                        last_updated: '2025-10-12T10:43:00+05:30'
                        leaders:
                        - symbol: TCS
                          name: Tata Consultancy Services
                          last_price: 3725.5
                          change_1d: 2.1
                          change_1w: 4.3
                          change_1m: 9.6
                          momentum_score: 0.88
                          turnover_ratio: 1.8
                          weight_pct: 24.3
                      leaders:
                      - symbol: INFY
                        name: Infosys
                        last_price: 1498.4
                        change_1d: 2.0
                        change_1w: 3.1
                        change_1m: 8.0
                        momentum_score: 0.79
                        turnover_ratio: 1.6
                        weight_pct: 18.7
                      laggards:
                      - symbol: LTI
                        name: LTIMindtree
                        last_price: 5550.1
                        change_1d: -0.6
                        change_1w: 0.2
                        change_1m: 3.4
                        momentum_score: 0.41
                        turnover_ratio: 0.9
                        weight_pct: 6.2
                    notes:
                    - Sector momentum computed from NSE index series (price return %).
                    - Turnover ratio compares current traded value vs 20-day average.
        '304':
          description: Not Modified
components:
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 2000
        default: 100
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      schema:
        type: string
...
  headers:
    ETag:
      description: Entity tag of response payload
      schema:
        type: string
...
  schemas:
    Symbol:
      type: string
      description: Canonical uppercase ticker
      example: RELIANCE
    RunId:
      type: string
      description: UTC timestamp in YYYYMMDDHHmmss (14 digits)
      pattern: ^[0-9]{14}$
      example: '20250912153000'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
          nullable: true
        next_cursor:
          type: string
          nullable: true
...
    UniverseInstrument:
      type: object
      properties:
        symbol:
          $ref: '#/components/schemas/Symbol'
        name:
          type: string
        sector:
          type: string
        active:
          type: boolean
      required:
      - symbol
      - name
...
    AdvanceDeclineSummary:
      type: object
      description: Advancer/decliner counts within a basket.
      required: [advancers, decliners]
      properties:
        advancers:
          type: integer
          minimum: 0
        decliners:
          type: integer
          minimum: 0
        unchanged:
          type: integer
          minimum: 0
          nullable: true
    MomentumHeatmapNodeBase:
      type: object
      description: Shared momentum metrics for sectors and industries.
      required:
      - name
      - symbol
      - change_1d
      - change_1w
      - change_1m
      - turnover_ratio
      - last_updated
      properties:
        name:
          type: string
          example: NIFTY IT
        symbol:
          type: string
          example: NIFTYIT
        change_1d:
          type: number
          format: float
          description: Percentage change versus the previous close.
        change_1w:
          type: number
          format: float
          description: Percentage change over the trailing 5 trading days.
        change_1m:
          type: number
          format: float
          description: Percentage change over the trailing 21 trading days.
        momentum_score:
          type: number
          format: float
          minimum: -1
          maximum: 1
          nullable: true
          description: Normalized composite score combining price & breadth factors.
        turnover_ratio:
          type: number
          format: float
          minimum: 0
          description: Current traded value divided by the 20-session average.
        turnover_label:
          type: string
          nullable: true
          description: Preformatted turnover string (e.g., "1.4x").
        last_updated:
          type: string
          format: date-time
          description: IST timestamp when the last datapoint was observed.
        note:
          type: string
          nullable: true
    MomentumHeatmapConstituent:
      type: object
      required: [symbol, name, last_price, change_1d]
      properties:
        symbol:
          type: string
          example: INFY
        name:
          type: string
          example: Infosys
        last_price:
          type: number
          format: float
          description: Last traded price in INR.
        change_1d:
          type: number
          format: float
        change_1w:
          type: number
          format: float
          nullable: true
        change_1m:
          type: number
          format: float
          nullable: true
        momentum_score:
          type: number
          format: float
          nullable: true
        turnover_ratio:
          type: number
          format: float
          minimum: 0
          nullable: true
        avg_turnover_cr:
          type: number
          format: float
          nullable: true
          description: Average traded value (20-day) in INR crore.
        weight_pct:
          type: number
          format: float
          nullable: true
          description: Weight of this constituent inside its parent basket.
    MomentumHeatmapIndustry:
      allOf:
      - $ref: '#/components/schemas/MomentumHeatmapNodeBase'
      - type: object
        properties:
          level:
            type: string
            enum: [industry]
          parent_symbol:
            type: string
            description: NSE sector symbol this industry rolls into.
          total_turnover_cr:
            type: number
            format: float
            nullable: true
          leaders:
            type: array
            items:
              $ref: '#/components/schemas/MomentumHeatmapConstituent'
          laggards:
            type: array
            items:
              $ref: '#/components/schemas/MomentumHeatmapConstituent'
    MomentumHeatmapSector:
      allOf:
      - $ref: '#/components/schemas/MomentumHeatmapNodeBase'
      - type: object
        properties:
          level:
            type: string
            enum: [sector]
          total_turnover_cr:
            type: number
            format: float
            nullable: true
            description: NSE traded value for the index (INR crore).
          advance_decline:
            $ref: '#/components/schemas/AdvanceDeclineSummary'
          industries:
            type: array
            items:
              $ref: '#/components/schemas/MomentumHeatmapIndustry'
          leaders:
            type: array
            items:
              $ref: '#/components/schemas/MomentumHeatmapConstituent'
          laggards:
            type: array
            items:
              $ref: '#/components/schemas/MomentumHeatmapConstituent'
    MomentumHeatmapResponse:
      type: object
      required: [as_of, trade_date, session, sectors]
      properties:
        as_of:
          type: string
          format: date-time
          description: IST timestamp when this snapshot was computed.
        trade_date:
          type: string
          format: date
          description: Trading date aligned to NSE calendar.
        session:
          type: string
          enum: [pre, cash, post, closed]
          description: NSE market session corresponding to the data.
        run_id:
          $ref: '#/components/schemas/RunId'
        source:
          type: string
          example: nse
        latency_sec:
          type: integer
          minimum: 0
          description: Age in seconds of the newest upstream datapoint.
        sectors:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/MomentumHeatmapSector'
        notes:
          type: array
          items:
            type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
...
    DrawerDetail:
      type: object
      required:
        - drawer_contract_version
        - symbol
        - trading_day
        - header
        - sparkline
        - score_breakdown
        - action_block
        - meters
        - next_action
        - diagnostics
      properties:
        drawer_contract_version:
          type: string
          example: "1.0.0"
          description: Semantic version for the drawer contract.
        scoring_rules_version:
          type: string
          example: "scores_v2"
          description: Version tag emitted by scoring layer.
        symbol:
          type: string
          example: "RELIANCE.NS"
        trading_day:
          type: string
          format: date
          example: "2025-09-19"
          description: Latest completed trading day used for this view.
        intraday_numerator_used:
          type: boolean
          example: false
          description: True if any intraday values were used in composing this card.
        header:
          $ref: '#/components/schemas/DrawerHeader'
        sparkline:
          $ref: '#/components/schemas/DrawerSparkline'
        score_breakdown:
          $ref: '#/components/schemas/DrawerScoreBreakdown'
        position:
          $ref: '#/components/schemas/DrawerPosition'
        action_block:
          $ref: '#/components/schemas/DrawerActionBlock'
        meters:
          $ref: '#/components/schemas/DrawerMeters'
        next_action:
          $ref: '#/components/schemas/DrawerNextAction'
        alerts:
          $ref: '#/components/schemas/DrawerAlerts'
        diagnostics:
          $ref: '#/components/schemas/DrawerDiagnostics'
        news_recent_hours:
          type: integer
          description: Default lookback window (hours) for the right-drawer news tab.
          example: 168

        # -------- Backward-compatibility aliases (deprecated) --------
        run_id:
          type: string
          deprecated: true
        as_of:
          type: string
          format: date-time
          deprecated: true
        name:
          type: string
          deprecated: true
        sector:
          type: string
          deprecated: true
        price:
          type: number
          format: double
          deprecated: true
          description: Use header.price instead.
        pct_today:
          type: number
          format: double
          deprecated: true
          description: Use header.pct_1d instead.
        score:
          type: integer
          minimum: 0
          maximum: 100
          deprecated: true
          description: Use score_breakdown.score_total_0_100.
        indicators:
          type: object
          additionalProperties: true
          deprecated: true
        badges:
          type: array
          items:
            $ref: '#/components/schemas/LegacyBadge'
          deprecated: true
        method_pill:
          type: string
          deprecated: true
        alert_templates:
          type: array
          items:
            type: string
          deprecated: true
        channels:
          nullable: true
          deprecated: true
    DrawerHeader:
      type: object
      required: [name, price, pct_1d, badges]
      properties:
        name:
          type: string
          example: "Tata Consultancy Services"
        sector:
          type: string
          nullable: true
          example: "IT"
        price:
          type: number
          format: double
          example: 4521.5
        pct_1d:
          type: number
          format: double
          example: 2.4
        badges:
          type: array
          items:
            $ref: '#/components/schemas/Badge'
    Badge:
      type: object
      required: [category, label]
      properties:
        category:
          type: string
          enum: [BREAKOUT, MOMENTUM, WATCH, IGNORE,ACTION]
        label:
          type: string
          example: "🚀 Very High Breakout"
    LegacyBadge:
      type: object
      properties:
        code:
          type: string
        text:
          type: string
        key:
          nullable: true
        label:
          nullable: true
        color:
          nullable: true
    DrawerSparkline:
      type: object
      required: [prices_30d]
      properties:
        prices_30d:
          type: array
          items: { type: number, format: double }
          minItems: 1
        ema10_30d:
          type: array
          items: { type: number, format: double }
          description: Optional overlay; same length as prices_30d when present.
        dates_30d:
          type: array
          items:
            type: string
            format: date
          description: ISO trading dates aligned 1:1 with prices_30d.
    DrawerScoreBreakdown:
      type: object
      required: [score_total_0_100, rsi14, adx14, relvol20, trend_rank, breakout_quality]
      properties:
        score_total_0_100:
          type: integer
          minimum: 0
          maximum: 100
          description: Canonical score (Full when available; else normalized Basic).
        score_source:
          type: string
          enum: [full, basic_fallback]
        score_basic:
          type: integer
          minimum: 0
          maximum: 12
          nullable: true
        score_basic_normalized:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        rsi14:
          type: number
          format: double
        adx14:
          type: number
          format: double
        relvol20:
          type: number
          format: double
        trend_rank:
          type: integer
          minimum: 0
          maximum: 10
        breakout_quality:
          type: integer
          minimum: 0
          maximum: 10
        proximity_52w_high_pct:
          type: number
          format: double
          nullable: true
        pivot_clear_pct:
          type: number
          format: double
          nullable: true
        base_len_bars:
          type: integer
          nullable: true
        data_gaps:
          type: array
          items: { type: string }
          description: Missing inputs when basic_fallback is used.
        stale:
          type: boolean
          description: True when score_source is basic_fallback.
    DrawerPosition:
      type: object
      properties:
        entry_price_locked:
          type: number
          format: double
          nullable: true
        qty:
          type: integer
          nullable: true
    DrawerActionBlock:
      type: object
      required: [breakeven_state, euphoria_state]
      properties:
        stop_now:
          type: number
          format: double
          nullable: true
        stop_method:
          type: string
          enum: [ATRxK, EMA_tight]
          nullable: true
        exit_close_threshold:
          type: number
          format: double
          nullable: true
        breakeven_state:
          type: string
          enum: [Pending, Active]
        euphoria_state:
          type: string
          enum: [Off, On]
    DrawerMeters:
      type: object
      required: [risk, euphoria]
      properties:
        risk:
          $ref: '#/components/schemas/Meter'
        euphoria:
          $ref: '#/components/schemas/Meter'
    Meter:
      type: object
      required: [level, score_0_100, basis, thresholds]
      properties:
        level:
          type: string
          enum: [Low, Medium, High]
        score_0_100:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          description: Quantitative calibration for the meter; null if not computable.
        basis:
          type: object
          additionalProperties: true
          description: Raw inputs used to compute the meter (e.g., atr14_pct, rsi14, adx14).
        thresholds:
          type: object
          required: [low_lt, medium_gte, high_gte]
          properties:
            low_lt:
              type: integer
              example: 33
            medium_gte:
              type: integer
              example: 33
            high_gte:
              type: integer
              example: 66
    DrawerNextAction:
      type: object
      required: [code, text, refs]
      properties:
        code:
          type: string
          enum:
            - SELL_NOW
            - SELL_TOMORROW
            - HOLD
            - HOLD_BREAKEVEN
            - HOLD_TIGHT
            - BUY_BREAKOUT
            - BUY_PULLBACK
            - BUY_STARTER
            - WATCH
            - IGNORE
        text:
          type: string
          example: "Hold — Trail stop at ₹4382.0"
        reasons:
          type: array
          items: { type: string }
        refs:
          type: object
          additionalProperties: { type: number, format: double }
          description: Numeric references used in the text (e.g., S/E/pivot/entry/A/B).
    DrawerAlerts:
      type: object
      properties:
        suggestions:
          type: array
          items: { type: string }
    DrawerDiagnostics:
      type: object
      required: [reason, rules_version]
      properties:
        reason:
          type: string
          example: "Yes — 86/100. strong momentum, trend strong, pivot cleared, volume strong"
        reason_text:
          type: string
          nullable: true
          description: Optional natural-language explanation; may be same as reason.
        rules_version:
          type: string
          example: "scores_v2"
        blocked_reason:
          type: string
          nullable: true
...